make cmark-gfm usable with multithreading
where to start... so this library adds extension support to cmark but
it does so by creating global variables that are modified by parser.feed

using tsan I was able to find out the variables being modified and made
them threadlocal. that solves my usecase but it's still not correct if,
for example, you do evented I/O and try to feed two parsers in parallel
in the same thread.

the point is that this state belongs to the parser, so that different
parser instances don't override one another, but for some reason it's
a global variable instead